[FIX] pipeline optimization parts that were failing

+++ src/deepmol/pipeline_optimization/objective_wrapper.py
@@ -97,8 +97,13 @@ class ObjectiveTrainEval(Objective):
             trial_id = str(trial.number)
             path = os.path.join(self.save_dir, f'trial_{trial_id}')
             pipeline = Pipeline(steps=self.objective_steps(trial, **self.kwargs), path=path)
-            pipeline.fit(train_dataset)
+            if pipeline.steps[-1][1].__class__.__name__ == 'KerasModel':
+                pipeline.fit(train_dataset, validation_dataset=test_dataset)
+            else:
+                pipeline.fit(train_dataset)
             score = pipeline.evaluate(test_dataset, [self.metric])[0][self.metric.name]
+            if score is None:
+                score = float('-inf') if self.direction == 'maximize' else float('inf')
 
             best_scores = self.study.user_attrs['best_scores']
             min_score = min(best_scores.values()) if len(best_scores) > 0 else float('inf')
@@ -114,11 +119,13 @@ class ObjectiveTrainEval(Objective):
                     if self.direction == 'maximize':
                         min_score_id = min(best_scores, key=best_scores.get)
                         del best_scores[min_score_id]
-                        shutil.rmtree(os.path.join(self.save_dir, f'trial_{min_score_id}'))
+                        if os.path.exists(os.path.join(self.save_dir, f'trial_{min_score_id}')):
+                            shutil.rmtree(os.path.join(self.save_dir, f'trial_{min_score_id}'))
                     else:
                         max_score_id = max(best_scores, key=best_scores.get)
                         del best_scores[max_score_id]
-                        shutil.rmtree(os.path.join(self.save_dir, f'trial_{max_score_id}'))
+                        if os.path.exists(os.path.join(self.save_dir, f'trial_{max_score_id}')):
+                            shutil.rmtree(os.path.join(self.save_dir, f'trial_{max_score_id}'))
 
             self.study.set_user_attr('best_scores', best_scores)
             return score
